==> 2_DataPreprocessing.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime

#Get ETL_DATE for Incremental Ingestion (enabled if ETL_DATE set to 'CURRENT_DATE')
from dotenv import load_dotenv
from pathlib import Path
import os

dotenv_path = Path('db_credentials.env')
load_dotenv(dotenv_path=dotenv_path)

ETL_DATE = os.getenv('ETL_DATE')

# For Incremental Ingestion (enabled if ETL_DATE set to 'CURRENT_DATE')
if ETL_DATE == 'CURRENT_DATE':
    ETL_DATE = datetime.today().strftime('%Y%m%d')
else:
   ETL_DATE = '20250322'
    
print(ETL_DATE)

df_raw = pd.read_csv(f"data/bronze_{ETL_DATE}_Airline_Reviews_withImageUrls.csv")
print(len(df_raw))
df_raw.head()

#### Rename "Unnamed: 0" to "RowId"

df = df_raw.copy()
df.rename(columns={"Unnamed: 0": "RowId"}, inplace=True)

#### Format and Convert Dates

df['Review Date'] = pd.to_datetime(df['Review Date'], format="mixed")
df['Date Flown'] = pd.to_datetime(df['Date Flown'], format="mixed")
print(len(df))
df.head()

#### EDA & Check for Nulls (especially for needed key fields)

# Examine nulls (first pass)
df.info()

# Replace any empty string columns with actual nulls
df = df.replace(r'^\s+$', np.nan, regex=True)

#### Replace Overall_Rating 'n' with null, Cast as float

df[df['Overall_Rating'] == 'n']
df['Overall_Rating'] = df['Overall_Rating'].replace('n', np.nan).astype(float)

df[df['Overall_Rating'] == 'n']

#### Add Unique Id

def generate_hash(row):
    value = f"{row['Airline Name']}|{row['Review_Title']}|{row['Review Date']}"
    return hashlib.sha256(value.encode()).hexdigest()

df['Id'] = df.apply(generate_hash, axis=1)

# Re-examine nulls (after converting empty string to nulls)
df.info()

Key columns needed do not have nulls

#### Persist to CSV file

df.to_csv(f"data/silver_{ETL_DATE}_Airline_Reviews_Preprocessed.csv", index=False, date_format='%Y-%m-%d')
