==> 3a_LostLuggage_Labeling.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime
import os
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.neighbors import NearestNeighbors

#Get ETL_DATE for Incremental Ingestion (enabled if ETL_DATE set to 'CURRENT_DATE')
from dotenv import load_dotenv
from pathlib import Path
import os

dotenv_path = Path('db_credentials.env')
load_dotenv(dotenv_path=dotenv_path)

ETL_DATE = os.getenv('ETL_DATE')

# For Incremental Ingestion (enabled if ETL_DATE set to 'CURRENT_DATE')
if ETL_DATE == 'CURRENT_DATE':
    ETL_DATE = datetime.today().strftime('%Y%m%d')
else:
   ETL_DATE = '20250322'
    
print(ETL_DATE)

df = pd.read_csv(f"data/silver_{ETL_DATE}_Airline_Reviews_Sentiment.csv")
print(len(df))
df.head()

#### Load embeddings from previously generated .npz file

review_text_embeddings = None
review_categories_embeddings = None
EMBEDDING_BACKUP_FILE = f"data/{ETL_DATE}_Airline_Reviews_gte_small.npz"
if os.path.exists(EMBEDDING_BACKUP_FILE):
    # Load the .npz file
    airline_quality_embeddings = np.load(EMBEDDING_BACKUP_FILE)
    
    # List the variables stored in the file
    print(airline_quality_embeddings.files)
    
    # Access the individual arrays
    review_text_embeddings = airline_quality_embeddings['review_text_embeddings']
    review_categories_embeddings = airline_quality_embeddings['review_categories_embeddings']

    print("review_text_embeddings:", review_text_embeddings)
    print("review_categories_embeddings:", review_categories_embeddings)
    
    # Close the file after use
    airline_quality_embeddings.close()
else:
    airline_quality_embeddings = None
    
#airline_quality_embeddings

#review_text_embeddings[0:1]

#### Convert review text to embeddings using thenlper/gte-small

from sentence_transformers import SentenceTransformer

# To fix SettingWithCopyWarning: A value is trying to be set on a copy of a slice from a DataFrame.
df_embed = df[['Id', 'Review']].copy()

if review_text_embeddings is None:
    embedding_model = SentenceTransformer("thenlper/gte-small")
    review_text_embeddings = embedding_model.encode(df_embed['Review'])
else:
    review_text_embeddings = review_text_embeddings

review_text_embeddings[0:1]

df_embed['review_gte_small_embeddings'] = review_text_embeddings.tolist()
print(len(df_embed))
df_embed.head()

#### Convert review category/ies to embeddings using thenlper/gte-small

review_categories = ['Lost Luggage']
    #, 'Cabin Crew Service', 'Ground Crew Service', 
    #'Seat Comfort', 'In-flight Amenities', 'Safety', 'Cleanliness',
    #'On-time Performance', 'Value for money']

df_review_categories = pd.DataFrame(review_categories, columns=['review_category'])
print(len(df_review_categories))
df_review_categories.head()

if review_categories_embeddings is None:
    embedding_model = SentenceTransformer("thenlper/gte-small")
    review_categories_embeddings = embedding_model.encode(df_review_categories['review_category'])
else:
    review_categories_embeddings = review_categories_embeddings
    
review_categories_embeddings[0:1]

df_review_categories['review_category_gte_small_embeddings'] = review_categories_embeddings.tolist()
df_review_categories[['review_category', 'review_category_gte_small_embeddings']]

#### Double check review category embeddings mapping

embedding_model = SentenceTransformer("thenlper/gte-small")
embedding_model.encode(df_review_categories['review_category'][0]).tolist()

df_review_categories['review_category_gte_small_embeddings'][0]

#import numpy as np

# Double check review category embeddings mapping
assert np.allclose(
    embedding_model.encode(df_review_categories['review_category'][0]).tolist(), 
    df_review_categories['review_category_gte_small_embeddings'][0],
    rtol=1e-3, atol=1e-4 #increase to allow greater deviation
    #rtol=1e-7, atol=1e-8 #decrease for stricter precision
), "Embeddings do not match within tolerance!"

#### Backup embeddings to a .npz file

np.savez(
    EMBEDDING_BACKUP_FILE,
    review_text_embeddings=review_text_embeddings,
    review_categories_embeddings=review_categories_embeddings
)

#### Compute cosine similarity for review category

review_category = 'Lost Luggage'
review_category_tag = 'lost_luggage'

#from sklearn.metrics.pairwise import cosine_similarity

category_embedding = df_review_categories[df_review_categories['review_category'] \
    == review_category]['review_category_gte_small_embeddings'].values[0]
#category_embedding

#df_embed['review_gte_small_embeddings'][0]

# Compute cosine similarity for each row
df_embed[f'{review_category_tag}_cosim'] = df_embed['review_gte_small_embeddings'].apply(
    lambda row_embedding: cosine_similarity(
        np.array(category_embedding).reshape(1, -1),
        np.array(row_embedding).reshape(1, -1)
    )[0, 0]
)

#### Compute kneighbors distance for review category

#from sklearn.neighbors import NearestNeighbors

# Function to normalize a vector (to remove difference in magnitude of phrase vs paragraph
def normalize_vector(vector):
    norm = np.linalg.norm(vector)  # Compute the norm (magnitude)
    if norm == 0:  # Avoid division by zero
        return vector
    return vector / norm

df_embed['review_gte_small_embeddings']

np.vstack(df_embed['review_gte_small_embeddings'].values)

normalize_vector(np.vstack(df_embed['review_gte_small_embeddings'].values))

# Need to be done only once
nn = NearestNeighbors(n_neighbors=len(df_embed), metric='minkowski')
nn.fit(np.vstack(df_embed['review_gte_small_embeddings'].values)) 
#nn.fit(normalize_vector(np.vstack(df_embed['review_gte_small_embeddings'].values))) 

distances, indices = nn.kneighbors(np.array(category_embedding).reshape(1, -1))
#distances, indices = nn.kneighbors(normalize_vector(np.array(category_embedding).reshape(1, -1)))
print(distances)
print(indices)

index_kndistance_mapping = dict(zip(indices[0], distances[0]))
#index_kndistance_mapping

# Assign kneighbors distances for each row
df_embed[f'{review_category_tag}_kndist'] = df_embed.index.map(index_kndistance_mapping)

#### Verify threshold for cosine similarity / kneighbors distance for review category

category_threshold = 0.84252

df_embed[df_embed[f'{review_category_tag}_cosim'] >= category_threshold]\
[['Id','Review',f'{review_category_tag}_cosim',f'{review_category_tag}_kndist']]\
    .sort_values(by=[f'{review_category_tag}_cosim'], ascending=True)
#0.843110 cutoff

# Random check a particular review
df_embed[df_embed.index == 12159]\
[['Id','Review',f'{review_category_tag}_cosim',f'{review_category_tag}_kndist']].to_string()
#false positives: 129, 18850, 7268

df_embed[df_embed[f'{review_category_tag}_kndist'] <= 0.560]\
[['Id','Review',f'{review_category_tag}_cosim',f'{review_category_tag}_kndist']]\
    .sort_values(by=[f'{review_category_tag}_kndist'], ascending=False)
#0.56015 cutoff (secondary metric only after cosim)

#### Add review_category_flag based on threshold

df_embed[f'is_{review_category_tag}_flag'] = df_embed[f'{review_category_tag}_cosim'] >= category_threshold

print(len(df_embed))
df_embed.head()

#### Merge back review text labels to original dataframe

df_merged = df.merge(df_embed, how='left', on='Id', suffixes=['','_embed']).drop(columns=['Review_embed'])
len(df_merged)
df_merged.head()

#### Persist to CSV file

df_merged.to_csv(f"data/silver_{ETL_DATE}_Airline_Reviews_LostLuggageLabel.csv", index=False, date_format='%Y-%m-%d')
