==> 5_UploadToDatabricksVolume.ipynb <==
from datetime import datetime

ETL_DATE = '20250322'

if ETL_DATE is None:
    ETL_DATE = datetime.today().strftime('%Y%m%d')

print(ETL_DATE)

#Get DB credentials
from dotenv import load_dotenv
from pathlib import Path
import os

dotenv_path = Path('db_credentials.env')
load_dotenv(dotenv_path=dotenv_path)

DATABRICKS_TOKEN = os.getenv('DATABRICKS_TOKEN')
DATABRICKS_INSTANCE = os.getenv('DATABRICKS_INSTANCE')
print(DATABRICKS_TOKEN)
print(DATABRICKS_INSTANCE)

#### THIS STEP IS AD-HOC ONLY: FULL EXPORT 20250328 to DATABRICKS

'''
#Get DB credentials
from dotenv import load_dotenv
from pathlib import Path
import os

dotenv_path = Path('db_credentials.env')
load_dotenv(dotenv_path=dotenv_path)

DB_HOST = os.getenv('DB_HOST')
DB_PORT = os.getenv('DB_PORT')
DB_NAME = os.getenv('DB_NAME')
DB_USER = os.getenv('DB_USER')
DB_PASSWORD = os.getenv('DB_PASSWORD')

-----------------------------
import psycopg2
import pandas as pd

# Connect to the PostgreSQL database
connection = psycopg2.connect(
    dbname=DB_NAME,
    user=DB_USER,
    password=DB_PASSWORD,
    host=DB_HOST,
    port=DB_PORT
)

# Define the SQL query
query = "SELECT * FROM silver_airline_quality_reviews;"

# Load the query result into a DataFrame
df_raw = pd.read_sql_query(query, connection)

# Close the connection
connection.close()

-----------------------------
print(len(df_raw))
df_raw

-----------------------------
ETL_DATE = '20250328'
df_raw.to_csv(f"data/silver_{ETL_DATE}_Airline_Reviews_LostLuggageLabel.csv", index=False, date_format='%Y-%m-%d')
'''

#### Upload Dataframe to Databricks volume

import requests

#Using Databricks Upload a file API
#https://docs.databricks.com/api/workspace/files/upload
databricks_instance = DATABRICKS_INSTANCE
local_path = f"data/silver_{ETL_DATE}_Airline_Reviews_LostLuggageLabel.csv"
remote_path = '/Volumes/workspace/data608/data608_volume/silver_airline_quality_reviews.csv'
headers = {'Authorization' : f"Bearer {DATABRICKS_TOKEN}"}
#print(headers)

with open(local_path, 'rb') as f:
    r = requests.put(
      f"{databricks_instance}/api/2.0/fs/files{remote_path}?overwrite=true",
      headers=headers,
      data=f)
    r.raise_for_status()

print(f"Uploaded '{local_path}' to '{remote_path}' on Databricks workspace {databricks_instance}")
