==> 8_TestChatbotAPI.ipynb <==
#Get DB credentials
from dotenv import load_dotenv
import os

from dotenv import load_dotenv
from pathlib import Path
import os

dotenv_path = Path('db_credentials.env')
load_dotenv(dotenv_path=dotenv_path)

DATABRICKS_TOKEN = os.getenv('DATABRICKS_TOKEN')
DATABRICKS_INSTANCE = os.getenv('DATABRICKS_INSTANCE')

import json
import requests
data = {
   "dataframe_split":{
      "columns":[
         "query"
      ],
      "data":[
         [
            #"Which airlines are notorious for 'lost luggage'?"
            #"Which airline has the best overall rating?"
            #"What is the most recent review based on 'review date'?"
            #"Which airline has the best overall rating?"
            #"Which airline has the worst overall rating?"
            #"Which airline and route has the worst overall rating?"
            #"Which airline has the most reviews of lost luggage?"
            #"Which airline has the most negative reviews?"
            #"Which airline has the least negative reviews?"
            #"Give an example of a bad review for the worst airline?"
            #"Give an example of a good review for the best airline?"
            #"How long is a flight from Calgary to Tokyo?"
            "Which airport has the worst reviews?"
            #"Which airline is the most expensive?"
            #"Which reviews has an image of a plane and provide the image url?"
            #"What are the common review image captions"
         ]
      ]
   }
}

headers = {"Context-Type": "text/json", "Authorization": f"Bearer {DATABRICKS_TOKEN}"}

response = requests.post(
    url=f"{DATABRICKS_INSTANCE}/serving-endpoints/data608-llm-chatbot/invocations", json=data, headers=headers
)

result = json.dumps(response.json()['predictions']).encode().decode('unicode_escape')
print(result.strip('["]').replace('\\"', '"'))
